Fix tests - Resolve json - Refactor module
