+++ tests/test_contributors.py
@@ -4,7 +4,7 @@ import unittest
 from unittest.mock import MagicMock, patch
 
 from contributors.contributor_stats import ContributorStats
-from contributors.contributors import get_all_contributors, get_contributors
+from contributors import get_all_contributors, get_contributors
 
 
 class TestContributors(unittest.TestCase):
@@ -28,12 +28,13 @@ class TestContributors(unittest.TestCase):
         get_contributors(mock_repo, "2022-01-01", "2022-12-31")
 
         mock_contributor_stats.assert_called_once_with(
-            "user",
-            False,
-            "https://avatars.githubusercontent.com/u/12345678?v=4",
-            100,
-            "https://github.com/owner/repo/commits?author=user&since=2022-01-01&until=2022-12-31",
-            "",
+            username='user', 
+            new_contributor=False, 
+            avatar_url='https://avatars.githubusercontent.com/u/12345678?v=4', 
+            contribution_count = 100,
+            commit_url ='https://github.com/owner/repo/commits?author=user&since=2022-01-01&until=2022-12-31', 
+            sponsor_info='', 
+            organisations=[]
         )
 
     @patch("contributors.get_contributors")
@@ -54,6 +55,7 @@ class TestContributors(unittest.TestCase):
                 100,
                 "commit_url",
                 "sponsor_url_1",
+                []
             ),
         ]
 
@@ -63,12 +65,13 @@ class TestContributors(unittest.TestCase):
             result,
             [
                 ContributorStats(
-                    "user",
-                    False,
-                    "https://avatars.githubusercontent.com/u/29484535?v=4",
-                    200,
-                    "commit_url, commit_url",
-                    "sponsor_url_1",
+                    username = "user",
+                    new_contributor=False,
+                    avatar_url = "https://avatars.githubusercontent.com/u/29484535?v=4",
+                    contribution_count=200,
+                    commit_url = "commit_url, commit_url",
+                    sponsor_info = "sponsor_url_1",
+                    organisations= [],
                 ),
             ],
         )
@@ -133,12 +136,13 @@ class TestContributors(unittest.TestCase):
 
         # Note that only user is returned and user2 is not returned here because there were no commits in the date range
         mock_contributor_stats.assert_called_once_with(
-            "user",
-            False,
-            "https://avatars.githubusercontent.com/u/12345678?v=4",
-            100,
-            "https://github.com/owner/repo/commits?author=user&since=2022-01-01&until=2022-12-31",
-            "",
+            username = "user",
+            new_contributor=False,
+            avatar_url="https://avatars.githubusercontent.com/u/12345678?v=4",
+            contribution_count=100,
+            commit_url="https://github.com/owner/repo/commits?author=user&since=2022-01-01&until=2022-12-31",
+            sponsor_info='',
+            organisations=[],
         )
